requirements.txt
Flask~=2.2.5
Bootstrap_Flask==2.2.0
Flask_CKEditor==0.4.6
Flask_WTF==1.2.0
WTForms~=3.0.1
clipspy~=1.0.2